# specs/001-ceos-financial-pwa/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CEOs Financial Management API
  description: API for the CEOs (מנכ"לים) personal finance PWA
  version: 1.0.0
  contact:
    name: CEOs Development Team
servers:
  - url: /api
    description: Next.js API routes
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the authenticated user's profile
      operationId: getCurrentUser
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # TRANSACTIONS
  # ============================================
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      description: Get paginated list of user's transactions with optional filters
      operationId: listTransactions
      security:
        - sessionCookie: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [INCOME, EXPENSE]
          description: Filter by transaction type
        - name: month
          in: query
          schema:
            type: string
            format: date
            example: '2026-01-01'
          description: Filter by month (first day of month)
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by category ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Items per page
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Transactions]
      summary: Create transaction
      description: Create a new income or expense transaction
      operationId: createTransaction
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction
      description: Get a single transaction by ID
      operationId: getTransaction
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Transactions]
      summary: Update transaction
      description: Update an existing transaction
      operationId: updateTransaction
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Transactions]
      summary: Delete transaction
      description: Permanently delete a transaction
      operationId: deleteTransaction
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Transaction deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HTTP-only session cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, createdAt]
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [INCOME, EXPENSE]
        amount:
          type: number
          format: decimal
          description: Amount in ILS
        date:
          type: string
          format: date
        categoryId:
          type: string
          format: uuid
          nullable: true
        categoryName:
          type: string
          nullable: true
          description: Denormalized category name for display
        paymentMethod:
          type: string
          enum: [CASH, CARD]
          nullable: true
        source:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, type, amount, date, createdAt, updatedAt]
    TransactionCreate:
      type: object
      properties:
        type:
          type: string
          enum: [INCOME, EXPENSE]
        amount:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 10000000
        date:
          type: string
          format: date
        categoryId:
          type: string
          format: uuid
          description: Required for EXPENSE
        paymentMethod:
          type: string
          enum: [CASH, CARD]
          description: Required for EXPENSE
        source:
          type: string
          maxLength: 100
          description: Required for INCOME
        description:
          type: string
          maxLength: 500
      required: [type, amount, date]
    TransactionUpdate:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 10000000
        date:
          type: string
          format: date
        categoryId:
          type: string
          format: uuid
        paymentMethod:
          type: string
          enum: [CASH, CARD]
        source:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required: [error, message]
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Authentication required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: Invalid request data
            details:
              amount: Amount must be greater than 0
...